Remove concept of static files
Now treats assembly payload file as any other virtual file, and does not append the encoded symbols file at all. To implement this, write-based code in Symbols was moved into AsmContext, and the assembly loader code is patched with the updated VROM address.
