+++ MMR.Randomizer/Builder.cs
@@ -3119,7 +3119,7 @@ namespace MMR.Randomizer
                 var fileIndex = RomUtils.AppendFile(extended.Bundle.GetFull());
                 var file = RomData.MMFileList[fileIndex];
                 var baseAddr = (uint)file.Addr;
-                asm.Symbols.WriteExtendedObjects(extended.GetAddresses(baseAddr));
+                asm.WriteExtendedObjects(extended.GetAddresses(baseAddr));
             }
 
             // Add extra messages to message table.
@@ -3274,14 +3274,11 @@ namespace MMR.Randomizer
             var originalMMFileList = RomData.MMFileList.Select(file => file.Clone()).ToList();
 
             byte[] hash;
-            AsmContext asm;
+            var asm = AsmContext.LoadInternal();
             if (!string.IsNullOrWhiteSpace(outputSettings.InputPatchFilename))
             {
                 progressReporter.ReportProgress(50, "Applying patch...");
                 hash = Patch.Patcher.ApplyPatch(outputSettings.InputPatchFilename);
-
-                // Parse Symbols data from the ROM (specific MMFile)
-                asm = AsmContext.LoadFromROM();
 
                 // Apply Asm configuration post-patch
                 WriteAsmConfigPostPatch(asm, hash);
@@ -3365,7 +3362,6 @@ namespace MMR.Randomizer
                 }
 
                 // Load Asm data from internal resource files and apply
-                asm = AsmContext.LoadInternal();
                 progressReporter.ReportProgress(71, "Writing ASM patch...");
                 WriteAsmPatch(asm);
 
@@ -3385,7 +3381,7 @@ namespace MMR.Randomizer
 
                 if (_randomized.Settings.DrawHash || outputSettings.GeneratePatch)
                 {
-                    var iconStripIcons = asm.Symbols.ReadHashIconsTable();
+                    var iconStripIcons = asm.ReadHashIconsTable();
                     OutputHashIcons(ImageUtils.GetIconIndices(hash).Select(index => iconStripIcons[index]), Path.ChangeExtension(outputSettings.OutputROMFilename, "png"));
                 }
             }
