+++ MMR.Randomizer/Models/Rom/MMFile.cs
@@ -11,7 +11,6 @@ namespace MMR.Randomizer.Models.Rom
         public byte[] Data;
         public bool IsCompressed;
         public bool WasEdited;
-        public bool IsStatic;
 
         public MMFile Clone()
         {
@@ -23,7 +22,6 @@ namespace MMR.Randomizer.Models.Rom
                 Cmp_End = Cmp_End,
                 IsCompressed = IsCompressed,
                 WasEdited = WasEdited,
-                IsStatic = IsStatic,
                 Data = Data?.ToArray()
             };
         }
