+++ assembly/src/Constants.asm
@@ -1,5 +1,5 @@
 ; Payload configuration constants
-.definelabel G_PAYLOAD_VROM,  0x03800000
+.definelabel G_PAYLOAD_VROM,  0x02EE8000
 .definelabel G_PAYLOAD_SIZE,  0x60000
 .definelabel G_PAYLOAD_END,   0x80780000
 .definelabel G_PAYLOAD_ADDR,  (G_PAYLOAD_END - G_PAYLOAD_SIZE)
